refactor: move set-birthday form submit event handler to separate file
Remove the event handler from index.js and add it to a new file to follow better separation of concerns.
Create a new file to handle the Interaction of modal submit for set user's birthday.
Improve code readability by using const and let appropriately, adding comments and removing redundant code.
